when running eval final get the meet of the synth function and top
